fix: resolved comments from code review.
