+++ Core/Core/en.lproj/Localizable.strings
@@ -194,3 +194,12 @@
 "TRACK_SELECTION.CARD.ACCESS_COURSE_DESCRIPTION" = "Limited access to non-graded activities and course material. Your progress and all access will be lost after %@";
 "TRACK_SELECTION.BUTTON.CONTINUE_WITH_FREE_TRACK" = "Continue with free track";
 "TRACK_SELECTION.BUTTON.CONTINUE_TO_PAYMENT" = "Continue to payment (%@)";
+
+"CERTIFICATE_VIEW.WATERMARK" = "PREVIEW";
+"CERTIFICATE_VIEW.TITLE" = "This is to certify that";
+"CERTIFICATE_VIEW.TITLE_COMPLETED" = "has successfully completed all courses and received passing grades for a Verified Certificate in";
+"CERTIFICATE_VIEW.COURSE_OFFERED" = "a course offered by Google, an online learning partnership between Google & edX.";
+"CERTIFICATE_VIEW.VERIFIED" = "Verified Certificate";
+"CERTIFICATE_VIEW.ISSUED" = "Issued August 2023";
+"CERTIFICATE_VIEW.VALIDITY_TITLE" = "Valid certificate ID:";
+"CERTIFICATE_VIEW.CERTIFICATE_ID" = "1234567890";
